perf(240): pack node geometry as Float32
Positions and orientations cross the worker boundary and sit in the
placement cache as Float32Array instead of Float64Array, halving each
retained geometry copy (113 MB to 56 MB per population at 4.7 M nodes)
and the transfer. Micron coordinates spanning ~10,000 um keep ~0.001 um
of resolution in f32, which is plenty for placement, picking and
electrode maths; every consumer already reads element-wise through
positionAt/placementAt and is dtype-agnostic.

+++ src/__tests__/circuit-nodes/populations-placement.test.tsx
@@ -108,7 +108,7 @@ const key = (population: NodePopulation) => nodesSessionKey('circuit-id', 'asset
 function geometry(count: number): NodeGeometry {
   return {
     count,
-    positions: new Float64Array(count * 3),
+    positions: new Float32Array(count * 3),
     orientations: null,
     morphologies: null,
   };
